Make SMTLibNamer obsolete
